adjust file structure to include screens components and navigation
