+++ .vscode/launch.json
@@ -1,0 +1,18 @@
+// {
+// 	// Use IntelliSense to learn about possible attributes.
+// 	// Hover to view descriptions of existing attributes.
+// 	// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+// 	"version": "0.2.0",
+// 	"configurations": [
+
+		
+
+
+// 		{
+// 			"name": "Attach to packager",
+// 			"cwd": "${workspaceFolder}",
+// 			"type": "reactnative",
+// 			"request": "attach"
+// 		}
+// 	]
+// }
