+++ package.json
@@ -18,7 +18,7 @@
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-native": "0.68.2",
-    "react-native-safe-area-context": "4.2.4",
+    "react-native-safe-area-context": "^4.2.4",
     "react-native-screens": "~3.11.1",
     "react-native-vector-icons": "^9.1.0",
     "react-native-web": "0.17.7"
