+++ screens/ProjectTimer.tsx
@@ -1,0 +1,12 @@
+import { Text, View } from 'react-native';
+import React from 'react'
+import { petsStyles } from '../styles';
+import { SafeAreaView } from 'react-native';
+
+export default function ProjectTimer () {
+	return(
+		<SafeAreaView>
+			
+		</SafeAreaView>
+	)
+}
