--- styles.js
@@ -1,22 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const primary= "#EA3323";
-const secondary= "#3E9F20";
-const accent= "#FFDBC6";
-const accent2= "#CDE6C5";
-const grey= "#AEAEAE";
-const white= "#FFFFFF";
-
-const styles = StyleSheet.create({
-    projectsPage: {
-      display: "flex",
-      alignItems: 'center',
-      justifyContent: 'center',
-      color: "red",
-      text1: {
-          color: primary,
-      }
-    }
-})
-
-export default styles;
